Memory optimisation

+++ include/mach/mach_aver_scctl.h
@@ -18,13 +18,13 @@ public:
     Brain.Screen.render(true, false);
   }
 
-  int setLinePrefix(int lineNumber, string text) {
+  short setLinePrefix(int lineNumber, string text) {
     Brain.Screen.clearLine(lineNumber, color::black);
     Brain.Screen.setCursor(lineNumber, 0);
     const char *toPrint = text.c_str();
     Brain.Screen.print(toPrint);
     Brain.Screen.render();
-    return text.length();
+    return (short) text.length();
   }
 
   void setValueOfLine(int lineNumber, int prefixLength, string text) {
